Add base model with action server
Very few intents, but enough examples to get it to recognize a name
entity within a question about whether Gandalf knows them.

The action server then grabs that name entity and determines if it
matches a Lord of the Rings friend name in lotr_friends.txt. The action
server returns an appropriate utterance based on the entity.